arreglado warning con conversion de int a float para obtener el width y el height del viewport arreglado bug a la hora de crear el vertextype con tangent

+++ MasterUAB/Graphics/ContextManager/ContextManager.h
@@ -104,8 +104,8 @@ public:
 
 	D3D11_VIEWPORT *getViewPort(){ return m_ViewPort; }
 	UAB_GET_PROPERTY(IDXGISwapChain*, SwapChain);
-	UAB_GET_PROPERTY(float, Width);
-	UAB_GET_PROPERTY(float, Height);
+	UAB_GET_PROPERTY(int, Width);
+	UAB_GET_PROPERTY(int, Height);
 };
 
 #endif
